@@ -10,7 +10,7 @@ license = "Apache License 2.0"
 [tool.poetry.dependencies]
 # The Python version here should be kept in sync with the one used in our
 # oldest tests in tox.ini.
-python = "3.7.*"
+python = "<3.8 >= 3.7"
 
 # Local dependencies
 # Any local packages that have dependencies on other local packages must be
@@ -61,6 +61,7 @@ google-api-python-client = "1.6.5"
 google-auth = "2.16.0"
 httplib2 = "0.9.2"
 idna = "2.6"
+importlib-resources = "1.3.1"
 ipaddress = "1.0.16"
 ndg-httpsclient = "0.3.2"
 parsedatetime = "2.4"
