@@ -1,6 +1,5 @@
 """Test for certbot_nginx._internal.configurator."""
 import sys
-import unittest
 from unittest import mock
 
 import OpenSSL
@@ -1075,22 +1074,21 @@ class InstallSslOptionsConfTest(util.NginxTest):
         file has been manually edited by the user, and will refuse to update it.
         This test ensures that all necessary hashes are present.
         """
-        import pkg_resources
-
+        if sys.version_info >= (3, 9):  # pragma: no cover
+            import importlib.resources as importlib_resources
+        else:  # pragma: no cover
+            import importlib_resources
+        
         from certbot_nginx._internal.constants import ALL_SSL_OPTIONS_HASHES
-        all_files = [
-            pkg_resources.resource_filename("certbot_nginx",
-                os.path.join("_internal", "tls_configs", x))
-            for x in ("options-ssl-nginx.conf",
-                      "options-ssl-nginx-old.conf",
-                      "options-ssl-nginx-tls12-only.conf")
-        ]
-        assert all_files
-        for one_file in all_files:
-            file_hash = crypto_util.sha256sum(one_file)
-            assert file_hash in ALL_SSL_OPTIONS_HASHES, \
-                f"Constants.ALL_SSL_OPTIONS_HASHES must be appended with the sha256 " \
-                f"hash of {one_file} when it is updated."
+
+        tls_configs_ref = importlib_resources.files("certbot_nginx").joinpath(
+            "_internal", "tls_configs")
+        with importlib_resources.as_file(tls_configs_ref) as tls_configs_dir:
+            for tls_config_file in os.listdir(tls_configs_dir):
+                file_hash = crypto_util.sha256sum(os.path.join(tls_configs_dir, tls_config_file))
+                assert file_hash in ALL_SSL_OPTIONS_HASHES, \
+                    f"Constants.ALL_SSL_OPTIONS_HASHES must be appended with the sha256 " \
+                    f"hash of {tls_config_file} when it is updated."
 
     def test_nginx_version_uses_correct_config(self):
         self.config.version = (1, 5, 8)
